Add a removeAll method to the ToastManager to allow dismissing all active toasts
This is useful for things like top level navigation or a big change in what is on screen where the toasts that might still be up pending a timer no longer make sense. If you save a resource and then navigate away to a totally different section of the app, the toast might be on top of something very different and weird. So, its handy to have an imperative way to just dismiss everything to make sure users don't get that dissonance.